+++ packages/node-client/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @stoker-platform/node-client
 
+## 0.5.28
+
+### Patch Changes
+
+- Updated dependencies
+    - @stoker-platform/types@0.5.16
+    - @stoker-platform/utils@0.5.22
+
 ## 0.5.27
 
 ### Patch Changes