+++ packages/system-functions/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @stoker-platform/system-functions
 
+## 0.5.31
+
+### Patch Changes
+
+- Updated dependencies
+    - @stoker-platform/types@0.5.16
+    - @stoker-platform/node-client@0.5.28
+    - @stoker-platform/utils@0.5.22
+
 ## 0.5.30
 
 ### Patch Changes