+++ packages/web-client/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @stoker-platform/web-client
 
+## 0.5.24
+
+### Patch Changes
+
+- Updated dependencies
+    - @stoker-platform/types@0.5.16
+    - @stoker-platform/utils@0.5.22
+
 ## 0.5.23
 
 ### Patch Changes